# api/openapi/public-api.yaml
openapi: 3.0.3
info:
  title: Public API
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/v1

paths:
  /auth/login:
    post:
      operationId: Login
      tags: [Auth]
      summary: Authenticate caller from Zitadel access token and create API session
      description: |
        Expects a Zitadel OAuth access token in the Authorization header.
        On success, returns an internal API session token for subsequent calls.
      security:
        - ZitadelBearer: []
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginPayload"
        "401":
          description: Unauthorized (missing/invalid Zitadel token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (user not provisioned locally or default-group membership missing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/switch-group:
    post:
      operationId: SwitchGroup
      tags: [Auth]
      summary: Switch active group for current session
      security:
        - SessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchGroupRequest"
      responses:
        "200":
          description: Active group switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwitchGroupPayload"
        "400":
          description: Bad request (invalid body/groupId)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing/invalid/expired session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (membership in target group required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/memberships:
    get:
      operationId: GetMemberships
      tags: [Auth]
      summary: List caller group memberships
      security:
        - SessionBearer: []
      responses:
        "200":
          description: Membership list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipsResponse"
        "401":
          description: Unauthorized (missing/invalid/expired session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects:
    get:
      operationId: ListProjects
      tags: [Projects]
      summary: List projects for active session group
      security:
        - SessionBearer: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectsResponse"
        "401":
          description: Unauthorized (missing/invalid/expired session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: CreateProject
      tags: [Projects]
      summary: Create project in active session group
      security:
        - SessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProjectPayload"
        "400":
          description: Bad request (invalid body/name/reference)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing/invalid/expired session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict (e.g. duplicate project constraints)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ZitadelBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT

    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: session-token

  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

    AuthContext:
      type: object
      required: [userId, email, displayName, activeGroupId]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        activeGroupId:
          type: string
          format: uuid

    LoginPayload:
      type: object
      required: [sessionToken, authContext]
      properties:
        sessionToken:
          type: string
        authContext:
          $ref: "#/components/schemas/AuthContext"

    SwitchGroupRequest:
      type: object
      required: [groupId]
      properties:
        groupId:
          type: string
          format: uuid

    SwitchGroupPayload:
      type: object
      required: [authContext]
      properties:
        authContext:
          $ref: "#/components/schemas/AuthContext"

    Membership:
      type: object
      required: [groupId, groupName, role, active]
      properties:
        groupId:
          type: string
          format: uuid
        groupName:
          type: string
        role:
          type: string
          enum: [owner, member]
        active:
          type: boolean

    MembershipsResponse:
      type: object
      required: [memberships]
      properties:
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/Membership"

    Project:
      type: object
      required: [id, groupId, name, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    ProjectsResponse:
      type: object
      required: [projects]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"

    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120

    CreateProjectPayload:
      type: object
      required: [project]
      properties:
        project:
          $ref: "#/components/schemas/Project"